rename: revert structure
